Refactoring of visual components
